[kahi_openalex_works] working in similarity for no doi works
