+++ Kahi_openalex_works/setup.py
@@ -83,7 +83,9 @@ def main():
             'kahi',
             'pymongo',
             'joblib',
-            'kahi_impactu_utils'
+            'kahi_impactu_utils',
+            'mohan',
+            'thefuzz'
         ],
     )
